+++ module-apps/application-phonebook/windows/PhonebookContact.hpp
@@ -12,9 +12,9 @@ static inline void fillContactData(std::string &data, ContactRecord &contact)
     utils::findAndReplaceAll(data, "$CONTACT_ALTERNATIVE_NAME$", contact.alternativeName);
     utils::findAndReplaceAll(data, "$CONTACT_NAME$", contact.getFormattedName(ContactRecord::NameFormatType::Title));
     utils::findAndReplaceAll(
-        data, "$CONTACT_NUMBER1$", (contact.numbers.size() == 1) ? contact.numbers[0].numberUser : "");
+        data, "$CONTACT_NUMBER1$", (contact.numbers.size() == 1) ? contact.numbers[0].number.getEntered() : "");
     utils::findAndReplaceAll(
-        data, "$CONTACT_NUMBER2$", (contact.numbers.size() == 2) ? contact.numbers[1].numberUser : "");
+        data, "$CONTACT_NUMBER2$", (contact.numbers.size() == 2) ? contact.numbers[1].number.getEntered() : "");
     utils::findAndReplaceAll(data, "$CONTACT_SPEED_DIAL$", contact.speeddial);
 }