+++ changelog.md
@@ -1,8 +1,11 @@
 # Mudita PurePhone changelog
 
-## [0.25.1] - 2020-06-??
+## Current release
 
 ### Added
+
+### Changed
+* `[phonebook]` Improved contact matching by phone number.
 
 ### Fixed
 * '[development]' Fixed colour accuracy in Linux simulator